:memo: :: Rename qrcode-scan module -> qrcode
