+++ .idea/gradle.xml
@@ -38,7 +38,7 @@
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/login" />
             <option value="$PROJECT_DIR$/feature/main" />
-            <option value="$PROJECT_DIR$/feature/qrcode-scan" />
+            <option value="$PROJECT_DIR$/feature/qrcode" />
             <option value="$PROJECT_DIR$/feature/sign-up" />
           </set>
         </option>
