+++ app/build.gradle.kts
@@ -24,7 +24,7 @@ dependencies {
     implementation(project(":feature:login"))
     implementation(project(":feature:sign-up"))
     implementation(project(":feature:main"))
-    implementation(project(":feature:qrcode-scan"))
+    implementation(project(":feature:qrcode"))
 
     implementation(libs.junit)
     androidTestImplementation(libs.androidx.test.ext)
