+++ app/src/main/java/com/goms/goms_android_v2/navigation/GomsNavHost.kt
@@ -18,8 +18,8 @@ import com.goms.main.navigation.navigateToOutingStatus
 import com.goms.main.navigation.navigateToStudentManagement
 import com.goms.main.navigation.outingStatusScreen
 import com.goms.main.navigation.studentManagementScreen
-import com.goms.qrcode_scan.navigation.navigateToQrScan
-import com.goms.qrcode_scan.navigation.qrcodeScanScreen
+import com.goms.qrcode.navigation.navigateToQrScan
+import com.goms.qrcode.navigation.qrcodeScanScreen
 import com.goms.sign_up.navigation.navigateToNumber
 import com.goms.sign_up.navigation.navigateToPassword
 import com.goms.sign_up.navigation.navigateToSignUp
