+++ feature/qrcode/.gitignore
