+++ feature/qrcode/proguard-rules.pro
