+++ feature/qrcode/src/main/AndroidManifest.xml
