+++ settings.gradle.kts
@@ -33,4 +33,4 @@ include(":core:common")
 include(":feature:login")
 include(":feature:sign-up")
 include(":feature:main")
-include(":feature:qrcode-scan")
+include(":feature:qrcode")
